requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.26.0
torch>=2.1.0
diffusers>=0.30.0
transformers>=4.43.0
accelerate>=0.30.0
safetensors>=0.4.5
